chore(client): Split into modules
