+++ cli/crates/client-generator/tests/ast.rs
@@ -1,0 +1,9 @@
+mod block;
+mod class;
+mod common;
+mod expression;
+mod function;
+mod import;
+mod interface;
+mod statement;
+mod types;
